Files renamed, code commented, boundary of volume changed
