+++ src/start_menu/main_settings/change_text_font.cpp
@@ -1,4 +1,4 @@
-#include "ChangeTextFont.h"
+#include "change_text_font.h"
 #include "../../utils/utils.h"
 #include <iostream>
