+++ src/start_menu/main_settings/change_text_size.cpp
@@ -1,4 +1,4 @@
-#include "ChangeTextSize.h"
+#include "change_text_size.h"
 #include "../../utils/utils.h"
 #include <iostream>
