+++ src/start_menu/main_settings/change_text_speed.cpp
@@ -1,9 +1,10 @@
-#include "ChangeTextSpeed.h"
+#include "change_text_speed.h"
 #include "../../utils/utils.h"
 #include <iostream>
 
 void ChangeTextSpeed()
 {
-	std::cout << "Give me the time you want between each rendered character (in ms): ";
+	std::cout << "Give me the time you want between each rendered character "
+				 "(in ms): ";
 	utils::GetUserInput(utils::g_sleep_for_ms);
 }
