+++ src/start_menu/main_settings/change_text_weight.cpp
@@ -1,4 +1,4 @@
-#include "ChangeTextWeight.h"
+#include "change_text_weight.h"
 #include "../../utils/utils.h"
 #include <iostream>
